Remove unused EXTV code, clean up code, pragma once around vpu.h
